keep every changed file within the 500-LoC cap
The CI cap (file-size-cap.yml) would have failed this PR. Split the new
runtime modules and test support into siblings:

- graph_runtime.rs: CaptureGuard -> graph_runtime/capture_guard.rs
- graph_runtime/types.rs: GraphFallbackReason -> graph_runtime/fallback.rs
- graph_runtime/tests.rs: metrics/churn/key tests -> tests_metrics.rs
- dflash_head/free_state_tests.rs: runtime helper -> free_state_support.rs

Allow-list four files that cannot be split cleanly, each with a rationale:
gpu.rs and gpu_impl.rs are a single trait / trait impl (E0119 forbids
spanning files), and decode_a.rs / verify_b.rs are each one long dispatch
whose graph-critical launch order must not fragment.
