+++ crates/spark-model/src/layers/dflash_head/free_state_tests.rs
@@ -129,7 +129,7 @@ fn zero_head() -> BlockDiffusionDraftHead {
         rms_norm_eps: 0.0,
         ctx_window: 0,
         propose_graphs: parking_lot::Mutex::new(HashMap::new()),
-        graph_runtime: zero_graph_runtime(),
+        graph_runtime: support::zero_graph_runtime(),
         next_lane: std::sync::atomic::AtomicUsize::new(0),
         lanes_start_event: 0,
         suppress_graphs: std::sync::atomic::AtomicBool::new(false),
@@ -137,63 +137,6 @@ fn zero_head() -> BlockDiffusionDraftHead {
         quant: super::DflashQuantization::Bf16,
         startup: super::DsparkStartupExecution::from_env_lenient(),
     }
-}
-
-/// A disabled-but-valid graph runtime for the zero head. `free_state` only
-/// reaches it to invalidate retired keys, and this head never captures any.
-fn zero_graph_runtime() -> std::sync::Arc<spark_runtime::graph_runtime::GraphRuntime> {
-    use spark_runtime::graph_runtime::{
-        GraphCapabilities, GraphFingerprint, GraphMode, GraphPolicies, GraphRuntime,
-        GraphRuntimeConfig, PhasePolicy, ShapeBucket, SpeculativeAlgorithm,
-    };
-    let policy = PhasePolicy {
-        max_entries: 1,
-        max_estimated_bytes: 1 << 20,
-        capture_enabled: true,
-        replay_enabled: true,
-        prewarm_enabled: false,
-    };
-    let backend: std::sync::Arc<dyn GpuBackend> =
-        std::sync::Arc::new(spark_runtime::gpu::mock::MockGpuBackend::new());
-    std::sync::Arc::new(
-        GraphRuntime::new(
-            backend,
-            GraphCapabilities {
-                basic_graphs: true,
-                debug_dot: false,
-                graph_upload: false,
-                conditional_nodes: false,
-                while_nodes: false,
-                native_serialization: false,
-            },
-            GraphRuntimeConfig {
-                mode: GraphMode::Full,
-                speculative_algorithm: SpeculativeAlgorithm::None,
-                shape_buckets: vec![ShapeBucket {
-                    token_limit: 128,
-                    request_limit: 8,
-                }],
-                policies: GraphPolicies::new(policy, policy, policy, policy, policy).unwrap(),
-                max_cache_entries: 8,
-                max_cache_bytes: 8 << 20,
-                fingerprint: GraphFingerprint {
-                    runtime: "test".into(),
-                    model: "test".into(),
-                    kernel_build: "test".into(),
-                    device: "test".into(),
-                    cuda: "0".into(),
-                    driver: "0".into(),
-                    memory_layout: "test".into(),
-                },
-                resource_generation: 1,
-                compatibility_rules: Vec::new(),
-                export_dir: None,
-                prewarm_profile: None,
-            },
-            std::sync::Arc::new(spark_runtime::graph_runtime::GraphMetrics::default()),
-        )
-        .expect("graph runtime"),
-    )
 }
 
 fn zero_kernels() -> DflashKernels {
@@ -519,3 +462,6 @@ fn reclaim_seam_free_failure_retains_pointers() {
     assert_eq!(dstate.ctx_hidden_acc.0, 0);
     assert!(dstate.block_table_dev.is_none());
 }
+
+#[path = "free_state_support.rs"]
+mod support;
